Any form of comment must now be accompanied by a postId, which can be used to summarize a user's comment or post.

+++ src/main/java/com/yugao/vo/post/CurrentPageVO.java
@@ -7,7 +7,7 @@ import java.util.List;
 @Data
 public class CurrentPageVO {
 
-    private Long currentPage;
+    private Integer currentPage;
 
     private List<CurrentPageItemVO> discussPosts;
